Fixes issue with tool table layout
The issue is that the .wrapper limits
the size of the page, but the contents
of the table are 100px too big.  So
I've changed the value of on-laptop
by increasing it by 100px.  This solves
the issue for normal displays, but
I suspect phones and tablets may
still be an issue.