Add new tool: Simulix (#79)
* Add new tool: Simulix

Simulix is a tool created to compile Simulink models into Functional Mockup Units. This is done by exporting the models source code and then generating the code needed to compile the model to the lib file needed. Currently only Co-simulation with the FMI version 2.0 is supported, where ME for FMI 2.0 is planned.

* Changed description of tool to include more information