New tool ISG-virtuos
ISG-virtuos is a realtime co-simulation platform for the Model-in-the-Loop, Software-in-the-Loop and Hardware-in-the-Loop simulation of digital twins for machines, plants and robots. Based on FMI, technology-specific simulation solutions can be integrated into the digital twin of a production plant. ISG-virtuos thus enables a comprehensive virtual commissioning of industrial control systems and illustrates interactions between process, machine, control system and user in one single integrated virtualization.