Add support for Google Analytics
There is a Window popping up before the main content informing the user
that cookies are used and that the preferences may be changed.
The content is viewable without setting this. The only annoying part
is that there is no button to disable all non-essential cookies, and
that the categories for functional cookies and ad cookies cannot be
disabled (we do not use these).

Once preferences have been set, they can be changed through a link at
the foot of the page.

TODO: A Google Analytics account needs to be created and maintained.
  This must not track more than the privacy policy states. The current
  code uses a placeholder XXXX for the Google analytics.
  Note when creating your own GA account for debugging that Chrome/Firefox
  incognito mode blocks GA (Chrome also does it by default if Do Not Track is
  set, and Firefox does it by default and you can only enable GA for the
  current session by clicking).
TODO: The privacy policy must be updated with the new tracking. It
  must state that:
  * cookies are used to collect data
  * the third-party Google Analytics is used to collect and process the data
  * what we do with the collected data (including who it is sent to)

When updating the consent you may need to refresh or click another link
before the tracking starts/stops.